tui: split menu.go clusters, migrate menu styles to styleSet (plan 023 F1 / C18)
Structural commit, no behavior change. The window-scroll state machine
moves to menu_window.go, project-name resolution to project_name.go, and
HitRect to hits.go. menuStep/menuStepClamp wrappers collapse into direct
menuStepDir calls, and menuRowHit.Cmd is dropped (row.Index >= 0 carries
the activatable bit). Menu bar and dropdown renderers now use styleSet
entries (MenuBarFill/MenuBrand/MenuHint/MenuCell/MenuCellHover and the
Dropdown* family) built once per theme in buildStyleSetFill with fixed
HeaderBG/BG surfaces — identical in the base and selection-band sets —
instead of per-frame inline lipgloss.NewStyle() allocations. Move
fidelity and per-style FG/BG/Bold values verified byte-identical against
HEAD.

Verified: make lint && make test && make test-race && make build.
